requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
openai>=1.51.0,<2.0.0
qdrant-client>=1.11.0,<2.0.0
PyMuPDF>=1.24.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
python-multipart>=0.0.9,<1.0.0
motor>=3.4.0,<4.0.0
pymongo>=4.7.0,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<5.0.0
